Move post collection button to the right
It's often happens to add something to the collection by accident,
depending on how you hold or navigate the phone, one handed, in bed or
whatever else, when you just want to hit that minus button.
To make sure it doesn't happen, move it to the right, to the follow button.
Which also totally makes sense.